i32: remove FLAC-specific kernels, keep generic int32 ops (#125)
The i32 package mixed generic integer SIMD primitives with a larger set that
encodes FLAC codec semantics (Rice cost search, fixed predictors, quantized-LPC
residual/restore, mid/side decorrelation). go-flac, the only consumer, has
vendored those kernels in-tree (tphakala/go-flac#27), so they no longer belong
in a general-purpose SIMD library.

Remove the FLAC-specific kernels and keep the genuinely generic int32 ops: Add,
Sub, Interleave2, Deinterleave2, MinMax. Stripped from the Go dispatch, the
pure-Go references, the amd64/arm64 .s TEXT blocks (each was a self-contained
block, no shared DATA/GLOBL), and the parity/fuzz/benchmark/example tests.
decorrelate.go (Add/Sub plus the removed MidSide) becomes the generic arith.go.

Promote the asm-verification harness from internal/asmencoding to a public
asmcheck package so downstream asm authors can share one harness instead of
duplicating it. Feature detection stays in simd/cpu.

Reword the i32, MinMax, and crc.Checksum16 docs (plus README and doc.go) to
drop the FLAC-codec framing: the CRC-16 0x8005 polynomial is shared by many
protocols (ARC, Modbus, USB), and the kept int32 ops are generic.

Verified: build, vet, and golangci-lint clean; go test ./... on amd64; the NEON
i32 kernels and PMULL crc on a Raspberry Pi 5 (arm64); and the arm64 WORD
encoding cross-check with SIMD_REQUIRE_OBJDUMP.

Closes #125
