+++ asmcheck/objdump_test.go
@@ -1,4 +1,4 @@
-package asmencoding
+package asmcheck
 
 import "testing"
