+++ i32/i32.go
@@ -1,9 +1,9 @@
 // Package i32 provides SIMD-accelerated operations on int32 slices.
 //
-// It is the integer counterpart to the f32/f64 packages and exists to serve
-// integer-domain DSP hot loops (for example a pure-Go FLAC codec) where the
-// per-sample work is integer arithmetic and channel (de)interleaving rather
-// than floating-point math.
+// It is the integer counterpart to the f32/f64 packages, covering the
+// element-wise integer arithmetic, signed min/max reduction, and channel
+// (de)interleaving that integer-domain DSP hot loops need where the per-sample
+// work is integer arithmetic rather than floating-point math.
 //
 // All functions automatically select the optimal implementation based on
 // runtime CPU feature detection and fall back to a pure-Go implementation on
